move rider related classes into model/rider; move the import related models into model/import
