+++ lib/widgets/pages/import_riders_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weforza/exceptions/exceptions.dart';
 import 'package:weforza/generated/l10n.dart';
-import 'package:weforza/model/import_riders_state.dart';
-import 'package:weforza/model/rider/import_riders_delegate.dart';
+import 'package:weforza/model/import/import_riders_delegate.dart';
+import 'package:weforza/model/import/import_riders_state.dart';
 import 'package:weforza/riverpod/file_handler_provider.dart';
 import 'package:weforza/riverpod/member/member_list_provider.dart';
 import 'package:weforza/riverpod/repository/serialize_riders_repository_provider.dart';
