+++ lib/widgets/pages/rider_form.dart
@@ -7,9 +7,9 @@ import 'package:weforza/exceptions/exceptions.dart';
 import 'package:weforza/generated/l10n.dart';
 import 'package:weforza/model/member.dart';
 import 'package:weforza/model/profile_image_picker_delegate.dart';
-import 'package:weforza/model/rider_form_delegate.dart';
-import 'package:weforza/model/rider_model.dart';
-import 'package:weforza/model/rider_validator.dart';
+import 'package:weforza/model/rider/rider_form_delegate.dart';
+import 'package:weforza/model/rider/rider_model.dart';
+import 'package:weforza/model/rider/rider_validator.dart';
 import 'package:weforza/riverpod/file_handler_provider.dart';
 import 'package:weforza/riverpod/member/member_list_provider.dart';
 import 'package:weforza/riverpod/member/selected_member_provider.dart';
